coupling organisation and user with Hibernate and H2

+++ server/src/main/java/urban/server/repositories/OrganisationRepository.java
@@ -9,7 +9,7 @@ public interface OrganisationRepository {
 
     void delete(Organisation user);
 
-    Organisation findById(int id);
+    Organisation findById(Long id);
 
     List<Organisation> findAll();
 }
